extract generation events into core microcrate
