+++ Cargo.toml
@@ -31,6 +31,7 @@ members = [
   "crates/bitnet-logits",        # SRP: pure logits transform functions
   "crates/bitnet-logits-filters", # SRP: reusable logits/probability filtering transforms
   "crates/bitnet-generation",    # SRP: decode-loop stopping logic & generation events
+  "crates/bitnet-generation-events-core", # SRP: reusable generation stream event/stat contracts
   "crates/bitnet-generation-stop-core", # SRP: reusable stop criteria/check core
   "crates/bitnet-engine-core",   # SRP: orchestration contracts (traits, session types)
   "crates/bitnet-repl-core",     # SRP: reusable chat REPL state/command primitives
@@ -130,6 +131,7 @@ default-members = [
   "crates/bitnet-logits",
   "crates/bitnet-logits-filters",
   "crates/bitnet-generation",
+  "crates/bitnet-generation-events-core",
   "crates/bitnet-generation-stop-core",
   "crates/bitnet-engine-core",
   "crates/bitnet-repl-core",
